add 4bit quant 16groups
